requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
cloudflare==4.0.0
Werkzeug==3.1.3